Fix rare dmarc dns query failure.
Case HB-8198

It seems it may be possible for the dnspython library to sometimes
return a DNS query in an unexpected data type. We not protect against
that and no longer try to parse the dns answer unless it is in the
expected form.

Changelog: Fix rare dmarc dns query failure.